requirements.txt
pydantic
fastapi
uvicorn

reportlab
PyPDF2
cryptography
requests
python-pkcs11
python-jose
passlib[bcrypt]

alembic
mysql-connector-python
sqlalchemy

python-dotenv
pymysql

python-multipart
python-dotenv
fpdf
twilio
jinja2

aiosqlite 
